Make text formatting consistent with other plugins
This is a minor update to the label (browserActionTitle) to make it consistent with most Firefox plugins. I have removed the period at the end of the label and switched the label to use Title Case (AP Style). This minor text formatting change will allow Form Filler to blend in better with other plugins in the Firefox overflow menu.